Have implemented drawables into mainview layout file
Is currently building and displaying the displayshoe.png in rows vertically down the screen. Am going to work on having a Top Picks button at the top, and underneath will be one big button for each brand of shoes which when pressed will take the user to the details view of the brand the user selected.